// frontend/src/types/avatarType.ts
export enum AvatarType {
  AVATAR_GREEN = 'AVATAR_GREEN',
  AVATAR_RED = 'AVATAR_RED',
  AVATAR_BROWN = 'AVATAR_BROWN',
  AVATAR_GRAY = 'AVATAR_GRAY',
  AVATAR_YELLOW = 'AVATAR_YELLOW',

  AVATAR_FAKER_GREEN = 'AVATAR_FAKER_GREEN',
  AVATAR_FAKER_RED = 'AVATAR_FAKER_RED',
  AVATAR_FAKER_BROWN = 'AVATAR_FAKER_BROWN',
  AVATAR_FAKER_GRAY = 'AVATAR_FAKER_GRAY',
  AVATAR_FAKER_YELLOW = 'AVATAR_FAKER_YELLOW',
}